Changes to resolve comments and package updates
