+++ react/example/package.json
@@ -10,7 +10,7 @@
     "build": "webpack --mode production"
   },
   "dependencies": {
-    "@egovernments/digit-ui-components": "0.2.0-beta.40",
+    "@egovernments/digit-ui-components": "0.2.0-beta.41",
     "@egovernments/digit-ui-libraries": "1.8.10",
     "@egovernments/digit-ui-module-common": "1.7.10",
     "@egovernments/digit-ui-module-core": "1.8.14",
