+++ react/modules/Project/package.json
@@ -19,7 +19,7 @@
     },
     "dependencies": {
         "@egovernments/digit-ui-react-components": "1.8.14",
-        "@egovernments/digit-ui-components": "0.2.0-beta.40",
+        "@egovernments/digit-ui-components": "0.2.0-beta.41",
         "lodash": "^4.17.21",
         "react": "17.0.2",
         "react-date-range": "^1.4.0",
