+++ react/modules/core/package.json
@@ -14,7 +14,7 @@
     "prepublish": "yarn build"
   },
   "dependencies": {
-    "@egovernments/digit-ui-components": "0.2.0-beta.40",
+    "@egovernments/digit-ui-components": "0.2.0-beta.41",
     "@egovernments/digit-ui-react-components": "1.8.14",
     "react": "17.0.2",
     "react-dom": "17.0.2",
