+++ react/modules/sample/src/pages/employee/index.js
@@ -14,6 +14,7 @@ import BoundarySearchComposer from "./BoundarySearchComposer";
 import TabSearch from "./TabSearch";
 import BoundarySearchComposerWithFilter from "./BoundarySearchComposerWithFilter";
 import NewSampleSearch from "./NewSampleSearch";
+import SampleSearchWithMultipleTabs from "./SampleSearchWithMultipleTabs";
 
 const SampleBreadCrumbs = ({ location }) => {
   const { t } = useTranslation();
@@ -68,9 +69,10 @@ const App = ({ path, stateCode, userType, tenants }) => {
         </React.Fragment>
         <PrivateRoute path={`${path}/components`} component={() => <Sample></Sample>} />
         <PrivateRoute path={`${path}/test`} component={() => <Loader></Loader>} />
-        <PrivateRoute path={`${path}/search`} component={() => <SampleSearch></SampleSearch>} />
+        <PrivateRoute path={`${path}/search-multiple-tabs`} component={() =><SampleSearchWithMultipleTabs/>} />
         <PrivateRoute path={`${path}/new-search`} component={() => <NewSampleSearch></NewSampleSearch>} />
         <PrivateRoute path={`${path}/searchComposer`} component={()=><BoundarySearchComposer></BoundarySearchComposer>}/>
+        <PrivateRoute path={`${path}/search`} component={() => <SampleSearch></SampleSearch>} />
         <PrivateRoute path={`${path}/searchComposerWithFilter`} component={()=><BoundarySearchComposerWithFilter></BoundarySearchComposerWithFilter>}/>
         <PrivateRoute path={`${path}/filter`} component={() => 
           <BoundaryFilter
