+++ react/ui-components/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [0.2.0-beta.40] - 2025-06-24
+### [0.2.0-beta.41] - 2025-06-27
 ## Enhancements
 - Fixed pagination issue in ResultsTableWrapper(for inboxSearchComposer)
